+++ src/ShortcutManager.js
@@ -1,0 +1,89 @@
+/** global: Garnish */
+/**
+ * Keyboard shortcut manager class
+ *
+ * This can be used to map keyboard events to the current UI "layer" (whether that's the base document,
+ * a modal, an HUD, or a menu).
+ */
+Garnish.ShortcutManager = Garnish.Base.extend(
+    {
+        shortcuts: null,
+        layer: 0,
+
+        init: function() {
+            this.shortcuts = [[]];
+            this.addListener(Garnish.$bod, 'keydown', 'triggerShortcut');
+        },
+
+        addLayer: function() {
+            this.layer++;
+            this.shortcuts.push([]);
+            return this;
+        },
+
+        removeLayer: function() {
+            if (this.layer === 0) {
+                throw 'Can’t remove the base layer.';
+            }
+            this.layer--;
+            this.shortcuts.pop();
+            return this;
+        },
+
+        registerShortcut: function(shortcut, callback) {
+            shortcut = this._normalizeShortcut(shortcut);
+            this.shortcuts[this.layer].push({
+                key: JSON.stringify(shortcut),
+                shortcut: shortcut,
+                callback: callback,
+            });
+            return this;
+        },
+
+        unregisterShortcut: function(shortcut) {
+            shortcut = this._normalizeShortcut(shortcut);
+            var key = JSON.stringify(shortcut);
+            for (var i = 0; i < this.shortcuts[this.layer].length; i++) {
+                if (this.shortcuts[this.layer][i].key === key) {
+                    this.shortcuts[this.layer].splice(i, 1);
+                    break;
+                }
+            }
+            return this;
+        },
+
+        _normalizeShortcut: function(shortcut) {
+            if (typeof shortcut === 'number') {
+                shortcut = {keyCode: shortcut};
+            }
+
+            if (typeof shortcut.keyCode !== 'number') {
+                throw 'Invalid shortcut';
+            }
+
+            return {
+                keyCode: shortcut.keyCode,
+                ctrl: !!shortcut.ctrl,
+                shift: !!shortcut.shift,
+            };
+        },
+
+        triggerShortcut: function(ev) {
+            var shortcut;
+            for (var i = 0; i < this.shortcuts[this.layer].length; i++) {
+                shortcut = this.shortcuts[this.layer][i].shortcut;
+                if (
+                    shortcut.keyCode === ev.keyCode &&
+                    shortcut.ctrl === Garnish.isCtrlKeyPressed(ev) &&
+                    shortcut.shift === ev.shiftKey
+                ) {
+                    ev.preventDefault();
+                    this.shortcuts[this.layer][i].callback(ev);
+                    break;
+                }
+            }
+        },
+    }
+);
+
+Garnish.shortcutManager = new Garnish.ShortcutManager();